Refactor app-frame integration
Integrate app-frame in _app instead of the mdx-docs and index components. The component app-drawer is part of the app-frame component. Before on each page load, a new drawer component got initiated. The state of the drawer was restored based on the current path. Even rendered on the server-side, this causes a flicker effect. Maybe excluding app-drawer from SSR can solve this issue. Further investigation is needed.