Resolve breaking changes with next and vercel
It seems vercel relies for a ton of tests upon next@canary. The vercel engineers release several versions of @vervel/next, which rely on the respective next canary version. This results in a non-deterministic state what about the distribution platform supports and what a forthcoming next version supports in terms of features and bug fixes.

I was removing and interactively assembling the project's pages folder—several conflicts were identified. The vercel deployment constantly terminated without a trace of the reason. A few days ago, the build succeeded with any error, and this with no relevant code changes applied. This took a long time to figure out!

https://github.com/vercel/vercel/issues/4758#issuecomment-656579986