Migrate from next-i18next to next-translate
From now on, this project uses the library next-translate in favor of next-i18next. The task to integrate translation should be easy, but it was challenging in the past, to say that. One primary argument of migration was the large dependency tree the current library has compared to other libraries. Often, those other libraries are not compatible with current deployment scenarios.

Increase the version of the nextJS dependency to the latest because of using a new feature called i18n routing.

Increase dependencies react and react-dom versions to the latest.
Integrate next-translate (experimental), remove next-i18next.

Modify nextJs configuration and adjust for the new translation approach, which uses a webpack loader internally.

Enable NextJs module alias feature

Migrate integration of translation - take1

Change the central _app file from a class component to a functional component.

Migrate integration of translation - take2

Adjust the integration in all necessary files of the project

Migrate integration of translation - take3

Relocate all translation resources from public/static/locales to the locales folder located at the project's root folder.

Some cleanup