Modification of the data model
One of the most significant changes makes operations to de-normalise data type-safe.
Other changes include the rebuilding of the data model including extensions of the generated data set.
In summary, the most important aspects are covered, which are assumed by the demonstrator of version 2

New implementation of the demonstrator

The changes made adapt the existing demonstrator to the new data model. Basically, these changes will be followed by further steps in the implementation, provided that the demonstrator is further expanded.

Add necessary dependencies