Refactor archer comps element, surface, container
Element: The component allows the passing of a render function (renderFn for now). If a renderFn gets passed, the function gets called gets called with both the ref, and dynamic references. Both references will be available in the component the render function contains in its functional body. If no render function gets passed, the passed component gets cloned, and the references get assigned respectively; see how cloneElement works.
ArcherSurface: Only small stylistic refactoring and null checks applied.
ArcherContainer: Provides two export variants, one standard component, second with forwardRef. To not pollute the function declaration of the component, the functional declaration gets generated with withForwardRef.

Cleanup ref and transition provider, reducer

The refs state contains a new reference called dynamicRef for now of type SelectHandles. The purpose is to instruct the component that lives as a child inside the component archer element.

Improve types

Control customBoxForward from the outside

The major change is the usage of React useImperativeHandle hook, which allows outside control of the component. In the future, this will get replaced by a callback reference, where a reference current is an object of functional references.

Adjust code for usage of comp customBoxForward