Control of structure and composition of seq diags
Auto Play component
The component enables the individual steps contained in a sequence diagram to be run through as a series. An individual animation time is assigned to each single step; once this time has elapsed, the animation jumps to the next single step.

Progressive Step Builder
Allows the interactive expansion of a sequence diagram by appending single steps to an existing sequence diagram at the end.