Make vercel build compatible with yarn berry
The option to set the node linker from "pnp" (default) to "node-modules" works, but this causes a different problem.

A fixed insert in the .yarnrc.yml file will cause the option to be set back and forth when using the plug and play mode on a local development machine.

The node-linker option "node-modules" should only be necessary when deploying on the vercel infrastructure.

Due to several tests performed on one project, it seems that a vercel deployment ignores all yarn install steps, such as pre-install before install. Setting the node-linker option at this point is not considered for a deployment! It seems that only the script "vercel-build" within a script of the same name defined in package.json can get overridden.

After a brief look into @vercel/next, it becomes clear that there is a way to send your install commands to the deployment process. The definition of these steps seems possible via the dashboard of a project, but I did not try this.

The following was tested using the vercel cli.

The advantage of the dynamic variant is that the node linker gets set when loading the source code into the infrastructure before initiating the dependencies' installation. In the example, an installation of yarn berry also gets executed.

Possible improvements:
In the future an environment variable will set the node linker like this YARN_NODE_LINKER="node-modules". Yarn will use that instead of whatever is in the configuration.