Onboard ws mdx compile and runtime
The modules added are similar to a library called next-mdx-remote. The main difference is implementing the logic in typescript and cleaner steps to generate the respective server and client code. The existing loading mechanism gets migrated from a webpack loader to an fs loading approach using dynamic pages in forthcoming commits.