Fix build, decrease MUI dependency version
Material UI v5.0.0-alpha.18 seems to be the last version before the emotion migration got started. We decided to wait for a more major version because investing in the required configuration steps is not affordable.
Next to applying version changes in dev or peer dependencies, the pinning is only successful when MUI internals, also provided as npm modules, get also fixed.

Add hyphening in Diagram; dedicated react components loaded into MDX require a separate declaration, the hyphening declaration in the component appContent is not enough.