Watch pages outside of nexts standard page folder
This repository uses dynamic pages. The standard files of Next, including a shell for dynamic pages, are located in the page folder next supports. The actual page content is located in a different folder /docs/src/pages. Allowing the development process to respond to file changes made, a signal from the filesystem is necessary whenever changes are made at those MDX files. It is not enough to enumerate folders in the webpack of next configuration. A library called "next-remote-watch" solves exactly this problem.