Make font in assets compatible with fontloader
Font loading in nextJs is complicated. A single misconfigured flag, e.g., priority, font loading, does not work or produces a wrong order in loading the font resulting in a bad result in a lighthouse performance test.

Currently, we rely on the library next-google-fonts, which handles all necessary configuration.

The react component-based assets in the project got generated using a library named "svgr." Back then, we did not know any better and defined a font family for the SVG elements in an asset.

Unfortunately, the font does not apply to the asset when loading with next-google-fonts, but it is the only option right now to get fast font loading across the application.

A temporary workaround removes the explicit font declaration on the few asset files integrated. In a future step, the assets have to get reexported with exactly this circumstance in mind.