Dynamic rendering on device spec only in pages
The previous commit mentioned two versions of nextJs 10.0.5-canary.9 and canary.8. The commit notes that the newer versions resolve the issue, but it does not. The reason is different and hard to debug.

I first thought of dynamically importing a component (nextJS feature) in which a conditional render gets applied, returning null for a specific (isMobile) condition instead of a JSX.Element produces the misbehavior, but I was also wrong.

It turns out that conditional rendering based on device characteristics (isMobile) only works at the page level instead of a component level, here the dynamic page slug.

- Remove device spec conditional render in appContentHeader, appTableOfContents
- Remove device spec conditional render in mdxDocs
- Introduce dedicated components of mdxDocs for desktop and mobile
- Add device spec conditional render in dynamic page (slug) rendering mdxDocs for desktop and mobile (also dynamic import those components)