Include locale in decision when to route to base
Fix changing locale with queries

Due to a change in route handling, the parameter "asPath" of a router push needs to be null to trigger a route change.

The occurs only when the existing path includes one or more query parameters. Changing the route from a default locale to a secondary locale was possible, but not the outer way around.

We have to monitor the situation in upcoming releases of NextJs; maybe a bug was introduced in the following commit, which deals with basePath with a query and evaluating the function parameters.

https://github.com/vercel/next.js/commit/47b7660aec950d2985f1d137d574fb84b935d555

It worked like a charm in 10.0.5-canary.9 but stopped working in 10.0.5-canary.10. From version 10.0.5-canary.10 the changes in this commit are necessary.

Only push on mount when query values are not empty

Update dependencies