Upgrade critical dependencies NextJs / Windups
Due to a breaking change in NextJS from version 10.0.6-canary.3 to 10.0.6-canary.4, the usual way of using the windups library is no longer possible. Some investigation shows that windups does not provide a commonJS output.

Note: In a simple NextJS application, loading windups is still possible. In a more complex setup, such as when using Yarn sub-modules as we use in the repository, this is no longer possible.

We have provided a refactoring of the building approach to the windups author. The following ticket and pull request tracks the progress.

https://github.com/sgwilym/windups/issues/10
https://github.com/sgwilym/windups/pull/11

For now, we roll a custom version of windups distributed to NPM, which provides those changes (@project-millipede/windups version 1.1.12)