Use default load behavior to load topics head
Somehow, loading modules with next/dynamic produces a wired and repeated nesting error in debug mode. Using the default behavior fixes the problem, but more places are affected by this. The rule of using dynamic import should be restricted if the component is lazily loadable or immediately required. In the case of the component topics-head, immediate loading gets needed.