Refactor structure loading of content components
Load MDX (mobile) document structure without SSR because it is unclear if the user is on a mobile device or not. Load atomic components such as h1 - h6 depend on whether they have to be interactive or not and the device used. Do component loading in both getContent / getStaticContentProps / getStaticProps respectively in the dynamic page render.