Erase device detect from getContent getStaticProps
As getStaticProps gets executed at build time, the server's device type is here real misleading as the purpose of this method is for data loading; as components are some data, this option does not work. Libraries such as next-mdx-remote avoid the definition of components in the compilation step. Further work is required to align our current approach with the approach of next-mdx-remote or use the library directly due to significant improvements in version 3.