Update MUI dependencies for V5 migration
The version of packages in use

- Core, Lab - 5.0.0-alpha.36
- Icons, Styles - 5.0.0-alpha.35

Remove the other MUI packages in the resolutions section of package.json. Previously the exact version of e.g. of the System package was required to define, e.g., Core 5.0.0-alpha.18 required System 5.0.0-alpha.18 but also not so proper pinning of versions of other MUI packages and version.