Return to previous font size in the toc-link comp
The body font size has changed from theme.typography.body2 (0.875rem) to theme.typography.body1 (1rem). To return to the previous size, we override the font size for the toc-link component.

Note: Applying the override in the theme also affects the use of the typography component with the default body1 variant.