Add style injection order, emotion before jss
The style library used by default in v5 is emotion. While migrating from JSS to emotion, and if you are using JSS style overrides for your components (for example, overrides created by makeStyles), we will need to take care of the CSS injection order. To do so, we need to have the StylesProvider with the injectFirst option at the top of your component tree.