Remove snapping of drawer animation
The mobile drawer width has to be permanently set to MAX_DRAWER_WIDTH; doing it only in the open state causes a fallback to width=100%, which is different from a specific with provided, resulting in a flickering when transitions from and to the state "open" occur. Remove conditional render for drawer closing.