Rewrite component app-frame
The application's layout was migrated from flex to grid. The reason for the migration was because a considerable content shift occurred.
Tests show that the layout grid is unnecessary to fix the content shift; in fact, the flex layout does not introduce any content shift once the refactorings are finished.
The flex version is still provided; minor adjustments are necessary to replace the grid with the flex-based version.