Resolve conflicting dependencies of ahooks
With the release of ahooks version 2.10.10 the build seems broken. It seems a conflict got introduced with core/js used in the use-request package, an internal module used in ahooks.
Pinning ahooks to version 2.10.9 does not resolve the problem; this is due to a new version of use-request with a minor version increase was also published. A workaround is to pin also the version number of use-request to the previous version, which is 2.8.10.