Avoid re-rendering on hash-change
Navigation to anchors should not cause any re-render; however, changing the hash through NextJs building blocks Link and Router does cause a re-render.

The solution is to memoize the Component using a subset of the properties of the router object (locale and slug) as dependables.

Other modifications move the analytics provider as the top parent component towards being a sibling next to the actual component.